requirements.txt
# App Requirements
# Django
Django==5.0.6
django-cors-headers==4.3.1
django-eventstream==5.1.0
django-filter==24.2
django-grip==3.5.0
django-redis==5.4.0
django-storages==1.14.3
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-extensions==3.2.3
drf-access-policy==1.5.0

# Database
psycopg2-binary==2.9.9
clickhouse-connect==0.7.8

# Application
jsonschema==4.22.0
pandas==2.2.2
numpy==1.26.4
pika==1.3.2
channels==4.1.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
redis==5.0.4

# Tasks & Background Jobs
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Server
gunicorn==22.0.0
uvicorn==0.30.1

# Utilities
PyJWT==2.8.0
pyparsing==3.1.2
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
setuptools==69.5.1
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
zstandard==0.22.0
